Fixed properties issue (#53)
1st issue wasn't an issue - multiple properties can appear in a single
toggleFrame, each one needs to be stopped to start a new one.

For the 2nd issue, windows didn't seem to be setting the actual height
to the same value as MAC - hence the height value was overriding the
r_height value.

Changed __dimensionWindow to pick the minimum of height & r_height when
no geom is specified.

Also fixed spelling mistake in r_height