Fixed bug in #83
Container config options weren’t being kept. Now, when you open a
container, the old config is used. This keeps row counts, sticky,
padding settings consistent.

Added dictionary of usedConfigs with type__title as a key

When a container is closed, the config is added to this dictionary.

When a container is opened again, the config is loaded from this
dictionary.